requirements.txt
appdirs==1.4.4
attrs==25.1.0
biopython==1.85
contourpy==1.3.0
cycler==0.12.1
cylp==0.92.3
dill==0.3.9
einops==0.8.1
exceptiongroup==1.2.2
fastobo==0.13.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h5py==3.13.0
hypothesis==6.127.2
importlib_resources==6.5.2
Jinja2==3.1.5
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.43.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
npysearch==1.3.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
pyteomics==4.7.5
python-dateutil==2.9.0.post0
pytz==2025.1
scipy==1.13.1
selfies==2.2.0
six==1.17.0
sortedcontainers==2.4.0
spectrum-utils==0.4.2
sympy==1.13.1
torch>=2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
zipp==3.21.0

pytorch-lightning~=2.5.0.post0
torchmetrics~=1.6.1
requests~=2.32.3